api/requirements.txt
# Flask dependencies
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
python-dotenv==1.0.1
ffmpeg-python==0.2.0
